% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incluir_codigo_ibge.R
\name{diagnostico_join}
\alias{diagnostico_join}
\title{Realizar diagnóstico do join}
\usage{
diagnostico_join(resultado)
}
\arguments{
\item{resultado}{Tibble que é resultado do left_join entre a tabela de referência
e a base de dados passada para a função \code{incluir_codigo_ibge()}}
}
\value{
A função imprime um diagnóstico na tela, informando se o join foi bem
sucedido ou não. Caso o join tenha falhado, a função imprime uma lista com os
municípios que não foram encontrados na tabela de referência.
}
\description{
Realizar diagnóstico do join
}
