% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incluir_codigo_ibge.R
\name{incluir_codigo_ibge}
\alias{incluir_codigo_ibge}
\title{Incluir o codigo do IBGE}
\usage{
incluir_codigo_ibge(
  data_set,
  tabela_referencia = munifacil::depara_muni_codigo(),
  diagnostico = TRUE
)
}
\arguments{
\item{data_set}{base de dados arrumada, com colunas muni_join e uf_join}

\item{tabela_referencia}{tabela de referência. Por padrão usa uma base interna}

\item{diagnostico}{imprimir diagnóstico?}
}
\description{
Incluir o codigo do IBGE
}
