% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar_colunas.R
\name{limpar_colunas}
\alias{limpar_colunas}
\title{Cria coluna limpa a partir do nome bruto dos municipios}
\usage{
limpar_colunas(data_set, col_muni, col_uf)
}
\arguments{
\item{data_set}{base de dados}

\item{col_muni}{coluna de municipio a ser arrumada}

\item{col_uf}{coluna de municipio a ser arrumada}
}
\description{
Cria coluna limpa a partir do nome bruto dos municipios
}
